Use OverloadedRecordDot for simple accessors
I found that OO style 'T.U.V' is far better than lens'
'T & U ^. V' as it is easier to read on the eyes and
projects cohesion and context tightly, for *SIMPLE*
nested record field access only.

Note that at the moment, we should avoid RecordUpdate extension.
We should *always* favor lenses for nested updates.

One major feature of lens is that we could pass variables to work with
lens, furthering potential of composability. Also, composite data
lookup seems to be pretty handy and might become a necessary
advantage as we work with increasingly, heavy nested
immutable structures.